+++ test/integration/tests/4488-newer-cabal-version/Main.hs
@@ -1,0 +1,6 @@
+import StackTest
+
+main :: IO ()
+main = do
+  stackErr ["--stack-yaml", "stack-bad.yaml", "build", "--dry-run"]
+  stack ["--stack-yaml", "stack-good.yaml", "build"]